requirements.txt
aioredis==1.2.0
appdirs==1.4.3
asgiref==3.2.1
asn1crypto==0.24.0
async-timeout==3.0.1
attrs==19.1.0
autobahn==19.8.1
Automat==0.7.0
Babel==2.7.0
beautifulsoup4==4.8.0
bs4==0.0.1
certifi==2019.6.16
cffi==1.12.3
channels==2.2.0
channels-redis==2.4.0
chardet==3.0.4
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
daphne==2.3.0
Django==2.2.4
django-ipware==2.1.0
django-phonenumber-field==3.0.1
fake-useragent==0.1.11
geoip2==2.9.0
hiredis==1.0.0
hyperlink==19.0.0
idna==2.8
incremental==17.5.0
lxml==4.4.1
maxminddb==1.4.1
msgpack==0.6.1
parse==1.12.1
phonenumbers==8.10.16
pkg-resources==0.0.0
pycparser==2.19
pyee==6.0.0
PyHamcrest==1.9.0
pyppeteer==0.0.25
pyquery==1.4.0
pytz==2019.2
requests==2.22.0
requests-html==0.10.0
six==1.12.0
soupsieve==1.9.3
sqlparse==0.3.0
tqdm==4.34.0
Twisted==19.7.0
txaio==18.8.1
urllib3==1.25.3
w3lib==1.21.0
websockets==8.0.2
zope.interface==4.6.0
